完善部门、分组信息管理

+++ src/main/java/com/zkxh/demo/controller/staff/StaffGroupController.java
@@ -31,7 +31,6 @@ public class StaffGroupController extends BaseController {
     @Resource
     private StaffGroupService staffGroupService;
 
-
     @PostMapping("staff/group/addGroup")
     @ApiOperation(value = "添加员工组别", notes = "添加员工的工作组")
     public String addStaffGroup(@RequestBody StaffGroupReqVo staffGroupReqVo) {
@@ -59,10 +58,10 @@ public class StaffGroupController extends BaseController {
     @Transactional
     @DeleteMapping("staff/group/deleteGroups")
     @ApiOperation(value = "删除员工组别", notes = "删除员工的工作组信息，支持批量删除")
-    public String deleteStaffGroup(@RequestParam List<Integer> ids) {
-
-        int len = ids.size();
-
+    public String deleteStaffGroup(@RequestParam Integer[] ids) {
+        int len = ids.length;
+        if (len < 0)
+            return ResultUtil.jsonToStringError(ResultEnum.REQUEST_DATA_IS_NULL);
         Integer res = staffGroupService.deleteStaffGroupsByGroupId(ids);
         if (res == len) {
             return ResultUtil.jsonToStringSuccess();
@@ -74,13 +73,27 @@ public class StaffGroupController extends BaseController {
 
     @GetMapping("staff/group/getGroupInfoByStaffDeptId")
     @ApiOperation(value = "通过员工所在的部门信息获取响应的所有的组别信息", notes = "通过的DeptId完成级联查询")
-    @ApiImplicitParam(name = "staffDeptId", value = "部门ID", required = true)
-    public String getStaffGroupByStaffDeptId(@RequestParam Integer staffDeptId) {
+    @ApiImplicitParam(name = "staffDeptId", value = "部门ID")
+    public String getStaffGroupByStaffDeptId(@RequestParam(name = "staffDeptId") Integer staffDeptId) {
+        System.out.println(staffDeptId);
         List<StaffGroupRespVo> res = staffGroupService.getStaffGroupByDeptId(staffDeptId);
         if (res.size() > 0) {
             return ResultUtil.jsonToStringSuccess(res);
         } else {
-            return ResultUtil.jsonToStringError(ResultEnum.FAILED);
+            return ResultUtil.jsonToStringError(ResultEnum.DATA_NOT_FOUND);
+        }
+    }
+
+
+    @GetMapping("staff/group/getAllGroupInfo")
+    @ApiOperation(value = "通过所有员工的组别信息", notes = "获取所有组别信息")
+    @ApiImplicitParam(name = "staffDeptId", value = "部门ID")
+    public String getAllGroupInfo() {
+        List<StaffGroupRespVo> res = staffGroupService.getAllStaffGroupInfo();
+        if (res.size() > 0) {
+            return ResultUtil.jsonToStringSuccess(res);
+        } else {
+            return ResultUtil.jsonToStringError(ResultEnum.DATA_NOT_FOUND);
         }
     }
