+++ src/main/java/com/zkxh/demo/service/staff/group/impl/StaffGroupServiceImpl.java
@@ -1,6 +1,10 @@
 package com.zkxh.demo.service.staff.group.impl;
 
+import com.zkxh.demo.common.enums.ResultEnum;
+import com.zkxh.demo.common.handle.RuntimeServiceException;
 import com.zkxh.demo.dao.staff.StaffGroupMapper;
+import com.zkxh.demo.dto.staff.StaffDto;
+import com.zkxh.demo.model.staff.Staff;
 import com.zkxh.demo.model.staff.StaffGroup;
 import com.zkxh.demo.model.staff.StaffGroupExample;
 import com.zkxh.demo.service.staff.group.StaffGroupService;
@@ -50,16 +54,19 @@ public class StaffGroupServiceImpl implements StaffGroupService {
     }
 
     @Override
-    public Integer deleteStaffGroupsByGroupId(List<Integer> ids) {
+    public Integer deleteStaffGroupsByGroupId(Integer[] ids) {
 
         int count = 0;
 
         for (Integer id : ids) {
+            Integer res = staffGroupMapper.countStaffNumByGroupId(id);
+            if (res > 0)
+                throw new RuntimeServiceException(ResultEnum.DATA_IS_NOT_NULL);
             staffGroupMapper.deleteByPrimaryKey(id);
             count++;
 
         }
-        if (count == ids.size()) {
+        if (count == ids.length) {
             return count;
         } else {
             return 0;
@@ -69,30 +76,46 @@ public class StaffGroupServiceImpl implements StaffGroupService {
     @Override
     public List<StaffGroupRespVo> getStaffGroupByDeptId(Integer staffDeptId) {
 
-        StaffGroup staffGroup = new StaffGroup();
-        staffGroup.setDeptId(staffDeptId);
-
-        StaffGroupExample staffGroupExample = new StaffGroupExample();
-        List<Integer> listOfstaffDeptId = new ArrayList();
-        staffGroupExample.createCriteria().andDeptIdIn(listOfstaffDeptId);
-
-        List<StaffGroup> staffGroupSList = staffGroupMapper.selectByExample(staffGroupExample);
+        List<StaffGroup> staffGroupSList = staffGroupMapper.selectStaffGroupsByDeptId(staffDeptId);
 
         StaffGroupRespVo staffGroupRespVo = null;
 
         List<StaffGroupRespVo> list = Collections.synchronizedList(new ArrayList<>());
-
-
         for (StaffGroup group : staffGroupSList) {
             staffGroupRespVo = new StaffGroupRespVo();
-
-            staffGroupRespVo.setDeptId(group.getDeptId());
             staffGroupRespVo.setGroupId(group.getGroupId());
             staffGroupRespVo.setGroupName(group.getGroupName());
+            staffGroupRespVo.setDeptName(group.getStaffDept().getDeptName());
+            staffGroupRespVo.setStaffList(group.getStaffList());
+            list.add(staffGroupRespVo);
+        }
+        return list;
+    }
 
+    @Override
+    public List<StaffGroupRespVo> getAllStaffGroupInfo() {
+        List<StaffGroup> staffGroupsList = staffGroupMapper.selectAllStaffGroups();
+        StaffGroupRespVo staffGroupRespVo = null;
+        List<StaffGroupRespVo> list = Collections.synchronizedList(new ArrayList<>());
+        for (StaffGroup group : staffGroupsList) {
+            staffGroupRespVo = new StaffGroupRespVo();
+            staffGroupRespVo.setGroupId(group.getGroupId());
+            staffGroupRespVo.setGroupName(group.getGroupName());
+            staffGroupRespVo.setDeptName(group.getStaffDept().getDeptName());
+            staffGroupRespVo.setStaffList(group.getStaffList());
             list.add(staffGroupRespVo);
         }
 
         return list;
+    }
+
+    @Override
+    public boolean checkGroupIsNotIncludeStaff(Integer groupId) {
+
+        Integer res = staffGroupMapper.countStaffNumByGroupId(groupId);
+        if (res > 0) {
+            return true;
+        }
+        return false;
     }
 }
